split InsertValuesToken to ShardingInsertValuesToken & EncryptInsertValuesToken
