+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/sql/token/pojo/generic/InsertValuesToken.java
@@ -19,10 +19,8 @@ package org.apache.shardingsphere.core.rewrite.sql.token.pojo.generic;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.RoutingUnitAware;
 import org.apache.shardingsphere.core.rewrite.sql.token.pojo.SQLToken;
 import org.apache.shardingsphere.core.rewrite.sql.token.pojo.Substitutable;
-import org.apache.shardingsphere.core.route.type.RoutingUnit;
 import org.apache.shardingsphere.core.rule.DataNode;
 import org.apache.shardingsphere.sql.parser.sql.segment.dml.expr.ExpressionSegment;
 import org.apache.shardingsphere.sql.parser.sql.segment.dml.expr.complex.ComplexExpressionSegment;
@@ -40,7 +38,7 @@ import java.util.List;
  * @author panjuan
  */
 @Getter
-public final class InsertValuesToken extends SQLToken implements Substitutable, RoutingUnitAware {
+public abstract class InsertValuesToken extends SQLToken implements Substitutable {
     
     private final int stopIndex;
     
@@ -50,44 +48,6 @@ public final class InsertValuesToken extends SQLToken implements Substitutable, 
         super(startIndex);
         this.stopIndex = stopIndex;
         insertValueTokens = new LinkedList<>();
-    }
-    
-    @Override
-    public String toString() {
-        StringBuilder result = new StringBuilder();
-        for (InsertValueToken each : insertValueTokens) {
-            result.append(each).append(", ");
-        }
-        result.delete(result.length() - 2, result.length());
-        return result.toString();
-    }
-    
-    @Override
-    public String toString(final RoutingUnit routingUnit) {
-        StringBuilder result = new StringBuilder();
-        appendInsertValue(routingUnit, result);
-        result.delete(result.length() - 2, result.length());
-        return result.toString();
-    }
-    
-    private void appendInsertValue(final RoutingUnit routingUnit, final StringBuilder stringBuilder) {
-        for (InsertValueToken each : insertValueTokens) {
-            if (isAppend(routingUnit, each)) {
-                stringBuilder.append(each).append(", ");
-            }
-        }
-    }
-    
-    private boolean isAppend(final RoutingUnit routingUnit, final InsertValueToken insertValueToken) {
-        if (insertValueToken.getDataNodes().isEmpty() || null == routingUnit) {
-            return true;
-        }
-        for (DataNode each : insertValueToken.getDataNodes()) {
-            if (routingUnit.getTableUnit(each.getDataSourceName(), each.getTableName()).isPresent()) {
-                return true;
-            }
-        }
-        return false;
     }
     
     @RequiredArgsConstructor
